// jolkr-app/src/api/generated/Message.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Attachment } from "./Attachment";
import type { MessageEmbed } from "./MessageEmbed";
import type { Reaction } from "./Reaction";
import type { JsonValue } from "./serde_json/JsonValue";

/**
 * Public message DTO.
 */
export type Message = { 
/**
 * Unique identifier.
 */
id: string, 
/**
 * Owning channel identifier.
 */
channel_id: string, 
/**
 * Author user identifier.
 */
author_id: string, 
/**
 * Message content (may be encrypted).
 */
content: string | null, 
/**
 * Encryption nonce when content is encrypted.
 */
nonce: string | null, 
/**
 * Whether the message has been edited.
 */
is_edited: boolean, 
/**
 * Whether the message is pinned.
 */
is_pinned: boolean, 
/**
 * Reply to identifier.
 */
reply_to_id: string | null, 
/**
 * Owning thread identifier.
 */
thread_id?: string, 
/**
 * Number of replies in this thread.
 */
thread_reply_count?: number | null, 
/**
 * Attached files.
 */
attachments: Array<Attachment>, 
/**
 * Aggregated reactions.
 */
reactions: Array<Reaction>, 
/**
 * Attached embeds.
 */
embeds: Array<MessageEmbed>, 
/**
 * Attached poll, if any.
 */
poll?: JsonValue, 
/**
 * Webhook identifier.
 */
webhook_id?: string, 
/**
 * Webhook name.
 */
webhook_name?: string, 
/**
 * Webhook avatar.
 */
webhook_avatar?: string, 
/**
 * Creation timestamp.
 */
created_at: string, 
/**
 * Last-update timestamp.
 */
updated_at: string, };